+++ apps/rest-api-server/src/chain.ts
@@ -1,0 +1,4 @@
+import { getChain } from "@blobscan/chains";
+import { env } from "@blobscan/env";
+
+export const chain = getChain(env.CHAIN_ID);