+++ packages/api/src/context.ts
@@ -9,8 +9,8 @@ import type {
 import type IORedis from "ioredis";
 
 import type { BlobPropagator } from "@blobscan/blob-propagator";
+import type { Chain } from "@blobscan/chains";
 import type { BlobscanPrismaClient } from "@blobscan/db";
-import type { Network } from "@blobscan/network-blob-config";
 
 import type { ApiClient } from "./utils";
 import { createApiClient } from "./utils";
@@ -22,7 +22,7 @@ export type CreateContextOptions =
 type CreateInnerContextOptions = Partial<CreateContextOptions> & {
   apiClient?: ApiClient;
   blobPropagator?: BlobPropagator;
-  network: Network;
+  chain: Chain;
   redis?: IORedis;
   prisma: BlobscanPrismaClient;
 };
@@ -45,7 +45,7 @@ export type ApiKeys = Partial<{
 export type CreateContextParams = {
   apiKeys?: ApiKeys;
   blobPropagator?: BlobPropagator;
-  network: Network;
+  chain: Chain;
   enableTracing?: boolean;
   prisma: BlobscanPrismaClient;
   redis?: IORedis;
@@ -53,7 +53,7 @@ export type CreateContextParams = {
 };
 
 export type TRPCInnerContext = {
-  network: Network;
+  chain: Chain;
   prisma: BlobscanPrismaClient;
   blobPropagator?: BlobPropagator;
   redis?: IORedis;
@@ -61,14 +61,14 @@ export type TRPCInnerContext = {
 };
 
 export function createTRPCInnerContext({
-  network,
+  chain,
   prisma,
   blobPropagator,
   apiClient,
   redis,
 }: CreateInnerContextOptions) {
   return {
-    network,
+    chain,
     prisma,
     blobPropagator,
     apiClient,
@@ -81,7 +81,7 @@ export type ContextScope = "web" | "rest-api";
 export function createTRPCContext({
   blobPropagator,
   prisma,
-  network,
+  chain,
   enableTracing,
   scope,
   apiKeys,
@@ -94,7 +94,7 @@ export function createTRPCContext({
         : undefined;
 
       const innerContext = createTRPCInnerContext({
-        network,
+        chain,
         prisma,
         apiClient,
         blobPropagator,