+++ packages/api/src/routers/blob/getCount.ts
@@ -76,8 +76,8 @@ export const getCount = publicProcedure
   .input(inputSchema)
   .use(withFilters)
   .output(outputSchema)
-  .query(async ({ ctx: { filters, prisma, network } }) => {
-    const blobsCount = await countBlobs(prisma, network.id, filters);
+  .query(async ({ ctx: { filters, prisma, chain } }) => {
+    const blobsCount = await countBlobs(prisma, chain.id, filters);
 
     return {
       totalBlobs: blobsCount,