+++ packages/chains/src/chains/index.ts
@@ -1,0 +1,5 @@
+export { gnosis } from "./gnosis";
+export { holesky } from "./holesky";
+export { hoodi } from "./hoodi";
+export { mainnet } from "./mainnet";
+export { sepolia } from "./sepolia";