+++ packages/db/prisma/seed/chain.ts
@@ -1,0 +1,3 @@
+import { getChain } from "@blobscan/chains";
+
+export const chain = getChain("mainnet");